requirements.txt
imutils==0.5.1
numpy==1.15.4
python-twitter==3.5
scikit-image==0.14.1
opencv-python==3.4.4.19
Flask==0.12.3
gunicorn==19.5.0
oauthlib==2.0.6
requests==2.9.1
requests-oauthlib==0.8.0
TwitterAPI==2.4.8
